requirements.txt
google-genai
pillow
httpx
requests
pydantic
